+++ src/main/java/com/artiffex/scm/web/businesstier/service/interfaz/CostoAcompanianteService.java
@@ -1,7 +1,9 @@
 package com.artiffex.scm.web.businesstier.service.interfaz;
 
+import java.math.BigDecimal;
+
 public interface CostoAcompanianteService {
 	
-	public float precioPorFecha();
+	public BigDecimal precioPorFecha();
 
 }
