+++ src/main/java/com/artiffex/scm/web/businesstier/service/interfaz/PaqueteHotelService.java
@@ -1,5 +1,6 @@
 package com.artiffex.scm.web.businesstier.service.interfaz;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.artiffex.scm.web.businesstier.utilidades.ComboSelect;
@@ -8,6 +9,6 @@ public interface PaqueteHotelService {
 	
 	public List<ComboSelect> listaComboSelect();
 	
-	public float precioPorPaquete(int idPaqueteHotel);
+	public BigDecimal precioPorPaquete(int idPaqueteHotel);
 
 }
